dx_kyouyuu/03_outputs/ai_news_save/requirements.txt
# いけともAIニュース保存用
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0
# 画像テキスト索引（OCR）用
Pillow>=9.0.0
pytesseract>=0.3.10
# token 自動更新 → GitHub Secret 更新用
PyNaCl>=1.5.0